Melhorias no regex, add nome, refatoração

+++ server-file/src/main/java/com/tiagods/prolink/model/Obrigacao.java
@@ -16,7 +16,7 @@ public class Obrigacao {
     private Obrigacao.Tipo tipo;
     private Year ano;
     private Month mes;
-    private int cliente;
+    private Long cliente;
     @PathLocation
     private String dirForJob;
